feat: rename prisma-postgres -> generator-prisma-client, rename .generated -> generated, adjust README, adjust imports
