+++ generator-prisma-client/nextjs-starter-webpack-with-middleware/.eslintrc.json
