+++ generator-prisma-client/nextjs-starter-webpack-with-middleware/README.md
@@ -28,14 +28,14 @@ To successfully run the project, you will need the following:
   ```prisma
   generator client {
     provider = "prisma-client"
-    output = "../lib/.generated/prisma"
+    output = "../lib/generated/prisma"
     previewFeatures = ["driverAdapters", "queryCompiler"]
     runtime = "nodejs"
   }
 
   generator edge {
     provider = "prisma-client"
-    output = "../lib/.generated/prisma-edge"
+    output = "../lib/generated/prisma-edge"
     runtime = "edge-light"
   }
   ```
@@ -48,7 +48,7 @@ Clone the repository, navigate into it and install dependencies:
 
 ```
 git clone git@github.com:prisma/prisma-examples.git --depth=1
-cd prisma-examples/prisma-postgres/nextjs-starter-webpack-with-middleware
+cd prisma-examples/generator-prisma-client/nextjs-starter-webpack-with-middleware
 npm install
 ```
 
@@ -74,7 +74,7 @@ DIRECT_URL="__YOUR_PRISMA_POSTGRES_DIRECT_CONNECTION_STRING__"
 NEXT_PUBLIC_URL="http://localhost:3000"
 ```
 
-Note that `__YOUR_PRISMA_POSTGRES_CONNECTION_STRING__` is a placeholder value that you need to replace with the values of your Prisma Postgres + Accelerate connection string. Notice that the Accelerate connection string has the following structure: `postgres://<username>:<password>@<host>:<port>/<database>?sslmode=require`.
+Note that `__YOUR_PRISMA_POSTGRES_CONNECTION_STRING__` is a placeholder value that you need to replace with the values of your Prisma Postgres + Accelerate connection string. Notice that the Accelerate connection string has the following structure: `prisma+postgres://accelerate.prisma-data.net/?api_key=<api_key_value>`.
 
 Note that `__YOUR_PRISMA_POSTGRES_DIRECT_CONNECTION_STRING__` is a placeholder value that you need to replace with the values of your Prisma Postgres direct TCP connection string. The direct connection string has the following structure: `postgres://<username>:<password>@<host>:<port>/<database>`.
