+++ generator-prisma-client/nextjs-starter-webpack-with-middleware/app/favicon.ico
