+++ generator-prisma-client/nextjs-starter-webpack-with-middleware/components/display-prisma-enums.tsx
@@ -1,8 +1,16 @@
 'use client'
 
-import { QuoteKind } from '@/lib/.generated/prisma/enums'
+import { QuoteKind } from '@/lib/generated/prisma/enums'
+import { useState, useEffect } from 'react'
 
 export function DisplayPrismaEnums() {
+  const [mounted, setMounted] = useState(false)
+
+  useEffect(() => {
+    // Note: this avoid "Hydration failed" errors in Next.js
+    setMounted(true)
+  }, [])
+
   return (
     <div className="w-full max-w-2xl mx-auto">
       <div className="border-4 border-dashed border-red-400 bg-yellow-50 dark:bg-yellow-900/30 rounded-lg p-4 text-xs text-yellow-800 dark:text-yellow-200 shadow-sm">
@@ -11,9 +19,7 @@ export function DisplayPrismaEnums() {
         </div>
         <div className="mb-1 text-center">
           Rendered on client?{' '}
-          <span className="font-mono">
-            {typeof window !== 'undefined' ? 'Yes' : 'No'}
-          </span>
+          <span className="font-mono">{mounted ? 'Yes' : 'No'}</span>
         </div>
         <div className="mb-2 text-center">
           <span className="font-semibold">
