+++ generator-prisma-client/nextjs-starter-webpack-with-middleware/components/quotes-loading.tsx
