+++ generator-prisma-client/nextjs-starter-webpack-with-middleware/lib/db.ts
@@ -1,5 +1,5 @@
 import { PrismaPg } from '@prisma/adapter-pg'
-import { PrismaClient } from './.generated/prisma/client'
+import { PrismaClient } from './generated/prisma/client'
 import { env } from './env/server'
 
 export type GetDbParams = {
