+++ generator-prisma-client/nextjs-starter-webpack-with-middleware/lib/prisma-enums.ts
@@ -1,0 +1,1 @@
+export * from './generated/prisma/enums'
