+++ generator-prisma-client/nextjs-starter-webpack-with-middleware/lib/types/index.ts
