+++ generator-prisma-client/nextjs-starter-webpack-with-middleware/lib/utils/helpers.ts
