+++ generator-prisma-client/nextjs-starter-webpack-with-middleware/next.config.mjs
