+++ generator-prisma-client/nextjs-starter-webpack-with-middleware/tailwind.config.ts
