+++ generator-prisma-client/nextjs-starter-webpack-with-middleware/tsconfig.json
