--- prisma-postgres/nextjs-starter-webpack-with-middleware/lib/prisma-enums.ts
@@ -1,1 +1,0 @@
-export * from './.generated/prisma/enums'
